Add files via upload
We're looking at six key variables to see how they influence home prices. Among the factors are the transaction date, house age, distance to the nearest MRT station, number of convenience stores nearby, latitude, and longitude. We're looking at this data to see if we can use it to estimate the house price. We will examine the correlation between the two variables, the R squared model, the f statistic, RSE, and P value, as well as create a linear regression model between a variable and the price of the house, in order to discover which of the variables does in fact influence the cost of the house. We'll also use the Q-Q plot and the plot for the density of the residual to assess the data's normality. R squared indicates how well the model fits the data and runs from 0 to 1, with 0 indicating that the model did not explain the outcome and 1 indicating that the linear regression model explained a significant percentage of the outcome. The Fstatistic indicates the model's overall significance, the higher the f statistic, the higher the model's quality. RSE aids in the calculation of the percentage of error (percent error = RSE/Y(Mean), with the lower the value, the better. Finally, suppose the p-value is less than 0.05. In that case, we should reject the null hypothesis, implying that the data is statistically significant, or accept it if the p-value is more than 0.05 and the data is not statistically significant. Based on all these factors, the third variable will be the best match for our data, as it has the highest R-squared of all the other models.